+++ src/oncoexporter/cda/cda_factory.py
@@ -1,4 +1,5 @@
 import abc
+import pandas as pd
 
 
 class CdaFactory(metaclass=abc.ABCMeta):
@@ -19,4 +20,22 @@ class CdaFactory(metaclass=abc.ABCMeta):
         for name in column_names:
             results.append(self.get_item(row, name))
         return results
+    
+    def days_to_iso(days: int):
+        """
+        Convert the number of days of life into an ISO 8601 period representing the age of an individual
+        (e.g., P42Y7M is 42 years and 7 months).
+
+        :param days: number of days of life (str or int)
+        """
+        if isinstance(days, str):
+            days = int(str)
+        if not isinstance(days, int):
+            raise ValueError(f"days argument must be int or str but was {type(days)}")
+        
+        # pandas does this conversion automatically: https://pandas.pydata.org/docs/reference/api/pandas.Timedelta.isoformat.html
+        td = pd.Timedelta(days=days)
+        iso = td.isoformat() # returns ISO 8601 duration string: td = pd.Timedelta(days=10350); td.isoformat(); 'P10350DT0H0M0S'
+        return iso
+
